restructured files; added css stylesheets
