+++ src/App.css
@@ -3,8 +3,4 @@
   margin: auto;
 }
 
-.hero-img{
-  height: 100vh;
-  background: url(./Images/Bonappetithero.jpg);
-  background-size: cover;
-}
+
