+++ src/Components/herostuff/Hero.jsx
@@ -1,8 +1,5 @@
 import React from "react";
-
-
-
-
+import 'hero.css';
 export default function Hero() {
   return (
     <div>
