+++ src/Components/herostuff/hero.css
@@ -1,0 +1,5 @@
+.hero-img{
+    height: 100vh;
+    background: url(./Images/Bonappetithero.jpg);
+    background-size: cover;
+  }
