+++ src/Pages/Contact.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Contact from '../Components/contact';
+import Contact from '../Components/contactstuff/contact';
 
 export default function Contact() {
   return (
